Align arm64_print declaration with uintptr_t definition

+++ app/src/main/cpp/box64/src/dynarec/arm64/arm64_printer.h
@@ -1,6 +1,9 @@
 #ifndef _ARM_PRINTER_H_
 #define _ARM_PRINTER_H_
 
-const char* arm64_print(uint32_t opcode, uint64_t addr);
+#include <stdint.h>
+#include <stddef.h>
+
+const char* arm64_print(uint32_t opcode, uintptr_t addr);
 
 #endif //_ARM_PRINTER_H_
